+++ core/src/com/mygdx/game/entities/Schmuck.java
@@ -224,7 +224,7 @@ public class Schmuck extends Entity implements Location<Vector2> {
 	 * @param y: y screen coordinate that represents where the tool is being directed.
 	 * @param wait: Should this tool wait for base cooldowns. No for special tools like built-in airblast/momentum freezing/some enemy attacks
 	 */
-	public void useToolStart(float delta, Equipment tool, short filter, int x, int y, boolean wait) {
+	public void useToolStart(float delta, Equipment tool, short filter, float x, float y, boolean wait) {
 		
 		//Only register the attempt if the user is not waiting on a tool's delay or cooldown. (or if tool ignores wait)
 		if ((shootCdCount < 0 && shootDelayCount < 0) || !wait) {
